fix: set finalizer to prevent leaks, always pass Regex by reference
Use of a finalizer mandates that a Regexp, once initialized, is never deep-copied, as the finalizer could then run on a previous copy of the Regexp, while the new copy is referencing the same internal pointers that may have been freed at that point. Therefore, prevent deep-copying by always taking references.